Merge front-end files previously in mygov-discovery repo into the mygov-bar repo.
This allows us to have the entire front end it one location, rather than split between two repos.

It also allows us to serve the entire front end as a single flat file.

In order to replace the asset pipeline, we'll run coffeescript and uglifyjs ourselves.

Note: css in /css, templates in /templates.