Allow parent page to pass settings to the child iframe.
The object MyGovConfig will be extended with defaults and included within MyGovLoader.config.

The Loader class will look for load = false to disable the load sequence.

Additionally, all config parameters will be passed via URL hash to the child iframe.

The child iframe (via crossdomain.coffee) will parse the parameter, and if whitelisted, store in MyGovBar.config.

This sounds complicated and anoying, but because the iframe and parent are on different domains, security restriction preclude them from talking to each other in any way other than postMessage.

This should work via postMessage with IE 8+ and with the hash hask with IE 7 and <.